@@ -30,28 +30,25 @@
         <skipTests>true</skipTests>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-        <spring.cloud.version>2021.0.1</spring.cloud.version>
-        <spring.cloud.alibaba.version>2021.0.1.0</spring.cloud.alibaba.version>
-        <mybatisplus.version>3.5.2</mybatisplus.version>
-        <knife4j.version>3.0.3</knife4j.version>
-        <springdoc.version>1.6.8</springdoc.version>
-        <hutool.version>5.7.22</hutool.version>
+        <java.version>17</java.version>
+        <spring.cloud.version>2022.0.0</spring.cloud.version>
+        <spring.cloud.alibaba.version>2022.0.0.0-RC1</spring.cloud.alibaba.version>
+        <mybatisplus.version>3.5.3</mybatisplus.version>
+        <knife4j.version>4.0.0</knife4j.version>
+        <hutool.version>5.8.11</hutool.version>
         <captcha.version>1.6.2</captcha.version>
-        <mapstruct.version>1.4.2.Final</mapstruct.version>
-        <nacos.version>2.1.1</nacos.version>
-        <hutool.version>5.7.22</hutool.version>
-        <aliyun.oss.version>3.8.0</aliyun.oss.version>
+        <mapstruct.version>1.5.3.Final</mapstruct.version>
+        <hutool.version>5.8.11</hutool.version>
+        <aliyun.oss.version>3.15.2</aliyun.oss.version>
         <dameng.version>8.1.2.79</dameng.version>
-        <aliyun.dysmsapi.version>2.0.18</aliyun.dysmsapi.version>
+        <aliyun.dysmsapi.version>2.0.23</aliyun.dysmsapi.version>
         <tencentcloud.sdk.version>3.1.574</tencentcloud.sdk.version>
-        <qiniu.version>7.11.0</qiniu.version>
-        <minio.version>8.4.3</minio.version>
-        <qcloud.cos.version>5.6.89</qcloud.cos.version>
+        <qiniu.version>7.12.1</qiniu.version>
+        <minio.version>8.5.1</minio.version>
+        <qcloud.cos.version>5.6.97</qcloud.cos.version>
         <huaweicloud.obs.version>3.22.3</huaweicloud.obs.version>
-        <oshi.version>6.2.2</oshi.version>
-        <easyexcel.version>3.1.1</easyexcel.version>
-        <easytrans.version>2.0.12</easytrans.version>
+        <easyexcel.version>3.1.5</easyexcel.version>
+        <easytrans.version>2.1.6</easytrans.version>
     </properties>
 
     <dependencies>
