@@ -3,7 +3,7 @@ package net.maku.message.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
-import net.maku.framework.common.page.PageResult;
+import net.maku.framework.common.utils.PageResult;
 import net.maku.framework.common.utils.ExceptionUtils;
 import net.maku.framework.common.utils.Result;
 import net.maku.message.convert.SmsPlatformConvert;
@@ -16,10 +16,11 @@ import net.maku.message.sms.service.SmsService;
 import net.maku.message.vo.SmsPlatformVO;
 import net.maku.message.vo.SmsSendVO;
 import org.apache.commons.lang3.StringUtils;
+import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -40,7 +41,7 @@ public class SmsPlatformController {
     @GetMapping("page")
     @Operation(summary = "分页")
     @PreAuthorize("hasAuthority('sms:platform:page')")
-    public Result<PageResult<SmsPlatformVO>> page(@Valid SmsPlatformQuery query){
+    public Result<PageResult<SmsPlatformVO>> page(@ParameterObject @Valid SmsPlatformQuery query){
         PageResult<SmsPlatformVO> page = smsPlatformService.page(query);
 
         return Result.ok(page);
