@@ -3,20 +3,21 @@ package net.maku.quartz.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
-import net.maku.framework.common.page.PageResult;
+import net.maku.framework.common.utils.PageResult;
 import net.maku.framework.common.utils.Result;
 import net.maku.quartz.convert.ScheduleJobLogConvert;
 import net.maku.quartz.entity.ScheduleJobLogEntity;
 import net.maku.quartz.query.ScheduleJobLogQuery;
 import net.maku.quartz.service.ScheduleJobLogService;
 import net.maku.quartz.vo.ScheduleJobLogVO;
+import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 /**
 * 定时任务日志
@@ -33,7 +34,7 @@ public class ScheduleJobLogController {
     @GetMapping("page")
     @Operation(summary = "分页")
     @PreAuthorize("hasAuthority('schedule:log')")
-    public Result<PageResult<ScheduleJobLogVO>> page(@Valid ScheduleJobLogQuery query){
+    public Result<PageResult<ScheduleJobLogVO>> page(@ParameterObject @Valid ScheduleJobLogQuery query){
         PageResult<ScheduleJobLogVO> page = scheduleJobLogService.page(query);
 
         return Result.ok(page);
