@@ -3,17 +3,18 @@ package net.maku.system.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
-import net.maku.framework.common.page.PageResult;
+import net.maku.framework.common.utils.PageResult;
 import net.maku.framework.common.utils.Result;
 import net.maku.system.convert.SysDictDataConvert;
 import net.maku.system.entity.SysDictDataEntity;
 import net.maku.system.service.SysDictDataService;
 import net.maku.system.query.SysDictDataQuery;
 import net.maku.system.vo.SysDictDataVO;
+import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.List;
 
 /**
@@ -31,7 +32,7 @@ public class SysDictDataController {
     @GetMapping("page")
     @Operation(summary = "分页")
     @PreAuthorize("hasAuthority('sys:dict:page')")
-    public Result<PageResult<SysDictDataVO>> page(@Valid SysDictDataQuery query){
+    public Result<PageResult<SysDictDataVO>> page(@ParameterObject @Valid SysDictDataQuery query){
         PageResult<SysDictDataVO> page = sysDictDataService.page(query);
 
         return Result.ok(page);
