@@ -3,7 +3,7 @@ package net.maku.system.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
-import net.maku.framework.common.page.PageResult;
+import net.maku.framework.common.utils.PageResult;
 import net.maku.framework.common.utils.Result;
 import net.maku.system.convert.SysDictTypeConvert;
 import net.maku.system.entity.SysDictTypeEntity;
@@ -11,10 +11,11 @@ import net.maku.system.service.SysDictTypeService;
 import net.maku.system.vo.SysDictVO;
 import net.maku.system.query.SysDictTypeQuery;
 import net.maku.system.vo.SysDictTypeVO;
+import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.List;
 
 /**
@@ -32,7 +33,7 @@ public class SysDictTypeController {
     @GetMapping("page")
     @Operation(summary = "分页")
     @PreAuthorize("hasAuthority('sys:dict:page')")
-    public Result<PageResult<SysDictTypeVO>> page(@Valid SysDictTypeQuery query){
+    public Result<PageResult<SysDictTypeVO>> page(@ParameterObject @Valid SysDictTypeQuery query){
         PageResult<SysDictTypeVO> page = sysDictTypeService.page(query);
 
         return Result.ok(page);
