@@ -3,18 +3,18 @@ package net.maku.system.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
-import net.maku.framework.common.page.PageResult;
+import net.maku.framework.common.utils.PageResult;
 import net.maku.framework.common.utils.Result;
 import net.maku.system.query.SysLogLoginQuery;
 import net.maku.system.service.SysLogLoginService;
 import net.maku.system.vo.SysLogLoginVO;
+import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.Valid;
-import java.util.Map;
+import jakarta.validation.Valid;
 
 /**
  * 登录日志
@@ -31,7 +31,7 @@ public class SysLogLoginController {
     @GetMapping("page")
     @Operation(summary = "分页")
     @PreAuthorize("hasAuthority('sys:log:login')")
-    public Result<PageResult<SysLogLoginVO>> page(@Valid SysLogLoginQuery query) {
+    public Result<PageResult<SysLogLoginVO>> page(@ParameterObject @Valid SysLogLoginQuery query) {
         PageResult<SysLogLoginVO> page = sysLogLoginService.page(query);
 
         return Result.ok(page);
