@@ -3,17 +3,18 @@ package net.maku.system.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
-import net.maku.framework.common.page.PageResult;
+import net.maku.framework.common.utils.PageResult;
 import net.maku.framework.common.utils.Result;
 import net.maku.system.convert.SysParamsConvert;
 import net.maku.system.entity.SysParamsEntity;
 import net.maku.system.query.SysParamsQuery;
 import net.maku.system.service.SysParamsService;
 import net.maku.system.vo.SysParamsVO;
+import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.List;
 
 /**
@@ -31,7 +32,7 @@ public class SysParamsController {
     @GetMapping("page")
     @Operation(summary = "分页")
     @PreAuthorize("hasAuthority('sys:params:all')")
-    public Result<PageResult<SysParamsVO>> page(@Valid SysParamsQuery query) {
+    public Result<PageResult<SysParamsVO>> page(@ParameterObject @Valid SysParamsQuery query) {
         PageResult<SysParamsVO> page = sysParamsService.page(query);
 
         return Result.ok(page);
