@@ -3,17 +3,18 @@ package net.maku.system.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
-import net.maku.framework.common.page.PageResult;
+import net.maku.framework.common.utils.PageResult;
 import net.maku.framework.common.utils.Result;
 import net.maku.system.convert.SysPostConvert;
 import net.maku.system.entity.SysPostEntity;
 import net.maku.system.service.SysPostService;
 import net.maku.system.query.SysPostQuery;
 import net.maku.system.vo.SysPostVO;
+import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.List;
 
 
@@ -32,7 +33,7 @@ public class SysPostController {
     @GetMapping("page")
     @Operation(summary = "分页")
     @PreAuthorize("hasAuthority('sys:post:page')")
-    public Result<PageResult<SysPostVO>> page(@Valid SysPostQuery query){
+    public Result<PageResult<SysPostVO>> page(@ParameterObject @Valid SysPostQuery query){
         PageResult<SysPostVO> page = sysPostService.page(query);
 
         return Result.ok(page);
