@@ -8,8 +8,8 @@ import net.maku.framework.common.utils.DateUtils;
 import net.maku.framework.common.utils.TreeNode;
 import org.hibernate.validator.constraints.Range;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
 import java.util.Date;
 
 /**
