@@ -7,8 +7,8 @@ import lombok.EqualsAndHashCode;
 import net.maku.framework.common.utils.DateUtils;
 import net.maku.framework.common.utils.TreeNode;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
 import java.util.Date;
 
 /**
