@@ -7,7 +7,7 @@ import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.info.License;
 import io.swagger.v3.oas.models.security.SecurityRequirement;
 import io.swagger.v3.oas.models.security.SecurityScheme;
-import org.springdoc.core.GroupedOpenApi;
+import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -37,7 +37,7 @@ public class SwaggerConfig{
             .title("MakuCloud")
             .description( "MakuCloud")
             .contact(contact)
-            .version("1.0")
+            .version("3.0")
             .termsOfService("https://maku.net")
             .license(new License().name("MIT")
             .url("https://maku.net")));
