@@ -9,7 +9,7 @@ import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.fhs.trans.service.impl.DictionaryTransService;
 import lombok.AllArgsConstructor;
 import net.maku.framework.common.exception.ServerException;
-import net.maku.framework.common.page.PageResult;
+import net.maku.framework.common.utils.PageResult;
 import net.maku.framework.mybatis.service.impl.BaseServiceImpl;
 import net.maku.system.convert.SysDictTypeConvert;
 import net.maku.system.dao.SysDictDataDao;
